langchain_trial/pyproject.toml
[project]
name = "langchain-trial"
version = "0.1.0"
description = ""
authors = [
    {name = "AlokGRetail",email = "alok.sg@gretail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "langchain-core (>=0.3.72,<0.4.0)",
    "langchain-openai (>=0.3.28,<0.4.0)",
    "openai (>=1.98.0,<2.0.0)",
    "langchain-anthropic (>=0.3.18,<0.4.0)",
    "google-generativeai (>=0.8.5,<0.9.0)",
    "langchain-huggingface (>=0.3.1,<0.4.0)",
    "transformers (>=4.54.1,<5.0.0)",
    "huggingface-hub (>=0.34.3,<0.35.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "numpy (>=2.3.2,<3.0.0)",
    "scikit-learn (>=1.7.1,<2.0.0)",
    "langchain (>=0.3.27,<0.4.0)",
    "langchain-community (>=0.3.27,<0.4.0)",
    "google (>=3.0.0,<4.0.0)",
    "google-genai (>=1.28.0,<2.0.0)",
    "torch (>=2.7.1,<3.0.0)",
    "accelerate (>=1.9.0,<2.0.0)",
    "sentence-transformers (>=5.0.0,<6.0.0)",
    "streamlit (>=1.47.1,<2.0.0)"
]

[tool.poetry]
packages = [{include = "langchain_trial", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
